WL-1398 Refactored the body of handleAccess() into 3 distinct methods.
The existing functionality already allows for output of the calendar as a
PDF file and as an iCal feed. Both of these hook into the handleAccess()
method, but as inline code, not distinct methods. So these each have their
own distinct method now, plus a placeholder one for handling opaque URLs.
Furthermore, due to an anticipated high degree of commonality between
handleAccessIcal() and handleAccessOpaqueUrl(), these in turn (will) share
a new common method - handleAccessIcalCommon().