WL-1398 Fleshed out the "controller" aspect of the opaque URL UI.
This commit very much builds on the previous one. In MVC terms, this
focussed on the VC and is intended to capture the essential logic of how
the UI switches in response to UI events in relation to the existence (or
otherwise) of an opaque URL. The M bit at present is implemented as a mock
persistence layer, which will not last beyond an application shutdown.