WL-1398 URL generation / handling talk to a unified mock persistence layer.
In previous commits, the parts that generated & presented the opaque URL to
the UI and the parts that handled an opaque URL, were created independently
and thus in fact used two separate mock persistence layers. This commit
encapsulates the first version that have both parts talking to the same
(mock) persistence layer. It has been implemented as a classic Sakai static
cover. Whilst several of the mock classes will disappear in subsequent
commits, others should follow through once the persistence is implemented
"for real" in Hibernate.