WL-1398 'Subscribe' link now appears for anyone with READ access.
In common with other permissions in CalendarAction, this is wrapped by a
method in an inner CalendarPermissions class. This new permission is called
'allowSubscribeThisCalendar' - subscribing to the implicit calendar (to
differentiate it from 'allowSubscribeCalendar - adding subscriptions to
external calendars). This is currently implemented as a call to the
calendar.read security function, i.e, no new security function was
introduced.