WL-1398 Reinstated and ensured current 'export' semantics are in effect.
Current 'export' semantics are:
- via the export route, you can only subscribe to an exported calendar.
- only the (typically auto-exported) calendar in MyWorkspace goes through
a READ authorization check.
On the other hand, access via opaque URL works independently of whether or
not it has been exported, i.e. it does not require it. It does however
*always* perform a READ authorization check, which means participants who
are removed from a site, will no longer be able to continue to access the
site calendar in this way. Thus the overall effect of this commit is that
backwards-compatibility is maintained, but site owners from here on in can
choose to not export the site calendar and instead point their users to the
opaque URL function instead.