WL-1398 Slight tweaks to improve error response sent back to client.
In a couple of places in the method that handles opaque URLs, it now throws
an EntityNotFoundException, in addition to logging the error. The reason
for this is that the corresponding catch clause that handles this will
effectively result in a 404 being sent back to the client. It's true that
most calendaring clients don't currently handle this is a very user visible
way, but they might in future and you can see this 'better' error handling
for yourself if you just type the opaque URL into the address bar of a web
browser. The reason it never throws a declared EntityPermissionException is
because this correspondingly results (due in part to the 'logout' in the
finally clause!) to result in a response giving the client an opportunity
to log in, which will only be offered by clients that support basic auth,
e.g. Thunderbird and it will not work anyway, even if "correct" credentials
are supplied (it probably doesn't logically make sense either!).