WL-1752 The UUID element of the opaque URL is now much more random.
The UUID element is now generated via UUID.randomUUID() which makes for a
much harder to guess UUID (it was using Hibernate's uuid generator class
before, which only guarantees 'uniqueness'). A highly visible knock-on
effect to the user is that when you regenerate an opaque URL, the UUID
component changes massively each time, which should lead to a much more
reassuring experience.