WL-1771 Accessing a private URL now respects any pre-existing session.
The modus operandi of handling a private URL is to implicitly create a
session by logging in as the user that the private UUID refers to,
outputting the calendar and then logging out the user. This works fine when
, for example, subscribed to a private URL in a calendaring client and
browsing WebLearn with a web-browser; each client access has its own
session and thus they never clash. If however, one was to access the
private URL with effectively the 'same client', then previously, the logout
that occurs at the end of the private URL would invalidate that common
session. The most obvious example of when this could occur is clicking on
the http(s) form of the private URL, when accessing WebLearn from a
web-browser. Thus the code now checks for any valid pre-existing session
and skips the (private URL access) logout if one is found.