WL-1464 Do some clean-up on event descriptions to produce valid iCal.
Calendar event descriptions that had line-breaks in them, seem to produce
invalid iCal. Unixen tools such as vi and less seem to render the
problematic character as ^M. When I ran the code in the Java debugger, the
actual char at that point in the string was '\r', so I just replace '\r'
with '\n', which seems to produce expected results. (Alas the iCalendar RFC
- RFC 5545 - was not a great deal of help in this regard).