WL-709 Added a 'Subscribe' link to the synoptic view of the calendar.
Clicking on the new 'Subscribe' link opens a page (within the synoptic
view) which contains explanatory text on how to subscribe via the main view
of the tool.

One change I had to make (that impacts on the main view) was I had to
change the names of message properties in order to cater for JSF EL syntax.
(What this meant in practice was changing the '.' character for '_'). This
enables each view to call upon the same set of underlying message
properties.

Conflicts:

	calendar-bundles/calendar.properties