WL-709 'Subscribe' in the main and synoptic views is now equivalent.
One change I had to make (that impacts on the main view) was I had to
change the names of message properties in order to cater for JSF EL syntax.
(What this meant in practice was changing the '.' character for '_'). This
enables each view to call upon the same set of underlying message
properties.