WL-3121 Don't blow up when the session is reset.
When the calendar tool loses it's session part way through it's request it can cause the managed bean to be created. The managed bean doesn't work and so the request blows up.

The workaround that the tool used was to initialise the resource loader using a JSP script-let at the top of the file. I've switched to creating a wrapper that allows the resource loader to be used as a managed bean. This is because managed beans are the standard way of working in the JSF world and it save the JSPs from having script-lets at the top of each file.